% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{getLogLine}
\alias{getLogLine}
\alias{getShortLogLine}
\alias{getLongLogLine}
\title{Log Line}
\usage{
getLogLine(line = c("1", "2", "3"), len = 10)

getShortLogLine(line = c("1", "2", "3"), len = 42L)

getLongLogLine(line = c("1", "2", "3"), len = 76L)
}
\arguments{
\item{line}{character vector, the type of line to return:
\describe{
\item{\code{"1"}}{character string with form '----------'}
\item{\code{"2"}}{character string with form '#--------#'}
\item{\code{"3"}}{character string with form '##########'}
}}

\item{len}{integer, length of the output line}
}
\value{
A character vector.
}
\description{
Returns a pre-defined character vector.
}
\section{Functions}{
\itemize{
\item \code{getShortLogLine()}: A character string with pre-defined length

\item \code{getLongLogLine()}: A character string with pre-defined length

}}
\author{
Alessandro Barberis
}
