% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{listAvailableLevels}
\alias{listAvailableLevels}
\title{List available levels}
\usage{
listAvailableLevels()
}
\arguments{
\item{path}{character string, a directory path}
}
\value{
A character vector containing the available logging levels.
}
\description{
This function lists the available logging levels.
See the \strong{Details} section below for further information.
}
\details{
Eight levels are supported:
\describe{
\item{\code{"OFF"}}{indicates the logger is inactive}
\item{\code{"FATAL"}}{indicates a fatal event in the application that prevents it from running}
\item{\code{"ERROR"}}{indicates an error in the application, possibly recoverable}
\item{\code{"WARN"}}{indicates an event that might lead to an error}
\item{\code{"INFO"}}{indicates informational messages}
\item{\code{"DEBUG"}}{indicates general debugging messages}
\item{\code{"TRACE"}}{indicates fine-grained debug events}
\item{\code{"ALL"}}{all levels are considered}
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
