% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{loggerConnectionUtilityFunctions}
\alias{loggerConnectionUtilityFunctions}
\alias{openCon}
\alias{closeCon}
\title{Connection Utility Functions}
\usage{
openCon(object)

closeCon(object)
}
\arguments{
\item{object}{an object of class \code{\link{Logger}}}
}
\description{
Functions to open and close connections.
}
\section{Functions}{
\itemize{
\item \code{openCon()}: Checks if \code{object} has a
connection or a path to a file. If so, returns a \code{\link{Logger}} with an
opened connection to a file.

\item \code{closeCon()}: Closes an opened connection.

}}
\examples{
logger = createLogger()

#open a connection
logger = openCon(logger)

#close a connection
closeCon(logger)


}
\author{
Alessandro Barberis
}
