% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{printLog}
\alias{printLog}
\title{Logger Print Functions}
\usage{
printLog(object, message, sep)
}
\arguments{
\item{object}{a \code{\link{Logger}}}

\item{message}{character string, the message to print}

\item{sep}{a character vector of strings to append after each element.}
}
\description{
This function prints the provided \code{message}
to the output defined via \code{object}.
}
\author{
Alessandro Barberis
}
\keyword{internal}
